+++ packages/@dopt/pkg/sync/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @dopt/pkg-sync
 
+## 1.1.2
+
+### Patch Changes
+
+- d405c1854: Add special case logic for copying the node and package manger version from the monorepo.
+
+## 1.1.1
+
+### Patch Changes
+
+- 459f0cb53: Add better error logging in the CLI.
+
 ## 1.1.0
 
 ### Minor Changes